re-organize authenticated file structure
