+++ src/authenticated/features/ChatScreen/styles.ts
