+++ src/authenticated/features/ChatScreen/useViewModel.ts
