+++ src/authenticated/features/Onboarding/styles.ts
