+++ src/authenticated/pages/Overview.tsx
@@ -1,7 +1,7 @@
 import ScreenLayout from '@components/ScreenLayout'
 import { Text } from 'react-native'
 import { BottomTabScreenProps } from '@react-navigation/bottom-tabs'
-import { RootAuthenticatedTabBarParamList } from '.'
+import { RootAuthenticatedTabBarParamList } from '..'
 
 type Props = BottomTabScreenProps<RootAuthenticatedTabBarParamList, '/overview'>
